Refactor farmer metrics.
